+++ src-tauri/src/database/mod.rs
@@ -1,4 +1,9 @@
 // Dir
+pub mod repositories {
+    pub mod mastery_repo;
+    pub mod task_repo;
+}
+
 pub mod services {
     pub mod item_services;
     pub mod migration_services;