+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be logged here
 - ✏️ move time store useEffect from App.tsx to AppInitializer.tsx
 - ✏️ add github actions to handle builds
 - ✏️ no longer supporting MSI installers
+- ✏️ refactor task_tracker and mastery_tracker commands seperating logic into repo's
 
 ## 1.1.0